Fix an issue where running `./gradlew clean assemble` would break the build
The root cause is that the `allowed-licenses.json` file is created at
the time the gradle build is evaluated. Post that, the `clean` task will
remove the `build` directory, and leaving no `allowedLicenses` file,
thereby causing the `checkLicense` task to fail.

Fix is to add an explicit dependency on a new task that ensures that
the `allowedLicenses` file exists, before `checkLicense` does its thing.